1、之前创建的工程师kotlin的，这里转为java         2、修改mp4v2的源码BUG，并添加一些调试日志

+++ app/src/main/cpp/libmp4v2-android/libutil/Utility.cpp
@@ -469,7 +469,7 @@ Utility::process_impl()
     formatGroups();
 
     // populate code lookup set
-    set<int> codes;
+    set<long long int> codes;
     const Group::List::const_iterator ie = _group.options.end();
     for( Group::List::const_iterator it = _group.options.begin(); it != ie; it++ ) {
         const Option& option = **it;
@@ -480,7 +480,7 @@ Utility::process_impl()
     }
 
     for( ;; ) {
-        const int code = prog::getOption( _argc, _argv, _shortOptions.c_str(), _longOptions, NULL );
+        const long long int code = prog::getOption( _argc, _argv, _shortOptions.c_str(), _longOptions, NULL );
         if( code == -1 )
             break;
